chore: enforce mypy --strict, improve test coverage
Fix mypy strict errors (explicit float casts for numpy returns),
remove numpy.mypy_plugin (unavailable), add tests for utility methods
and edge cases. mutation.py: 100%, store.py: 99%, metrics.py: 100%.

+++ src/gc_memory/store.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import math
 
-import faiss  # type: ignore[import-untyped]
+import faiss
 import numpy as np
 import numpy.typing as npt
 
@@ -125,7 +125,7 @@ class GCMemoryStore:
 
     def _get_sigma(self, affinity: float) -> float:
         """Compute mutation sigma. Override in baselines for fixed rate."""
-        return self.config.sigma_0 * (1.0 - affinity) ** self.config.gamma
+        return float(self.config.sigma_0 * (1.0 - affinity) ** self.config.gamma)
 
     def _mutate_gc_entries(
         self,
